fix: Scan timings
The esp scan behaves differently when connected to AP and disconnected
from AP. To prevent this issue we intentionally disconnect from the wifi
if connected and use the delayed disconnect handling (done in previous
commit) to reconnect after scan.

Also changed the waiting times of the scan to be more incremental.

These changes makes the results appear faster with more incremental
popins.

BFW-5387