esp: Use C++ instead of C
This is done mainly to make the new protocol extensions implementations
easier to port from the printer codebase, which is mainly c++. Also this
will allow us to use more safer interfaces. Also this commit only makes
the firmware compile, the codebase still contains a lot of C-isms that
could be replaced with better C++ alternatives.

BFW-5662
BFW-5659
BFW-5660
BFW-5609